REFACTOR : http request, response package 분리
