+++ src/main/java/Controller/Controller.java
@@ -1,7 +1,7 @@
 package Controller;
 
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 
 import java.io.IOException;
