+++ src/main/java/http/response/HttpResponse.java
@@ -1,5 +1,7 @@
-package http;
+package http.response;
 
+import http.HttpHeader;
+import http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
