+++ src/main/java/http/response/HttpStatusLine.java
@@ -1,4 +1,6 @@
-package http;
+package http.response;
+
+import http.HttpStatus;
 
 public class HttpStatusLine {
     HttpStatus httpStatus;
