+++ src/main/java/webserver/RequestHandler.java
@@ -1,8 +1,8 @@
 package webserver;
 
 import Controller.Controller;
-import http.HttpRequest;
-import http.HttpResponse;
+import http.request.HttpRequest;
+import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
